+++ backend/src/models/invoiceItem.model.js
@@ -8,15 +8,15 @@ module.exports = (sequelize) => {
       allowNull: false,
     },
     unit_price: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.FLOAT,
       allowNull: false,
     },
     tva: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.FLOAT,
       defaultValue: 0,
     },
     discount: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.FLOAT,
       defaultValue: 0,
     },
     total_item: {
